+++ app.js
@@ -1,11 +1,11 @@
 require('dotenv').config();
 const express = require('express');
 const mongoose = require('mongoose');
-const rateLimit = require('express-rate-limit');
 const helmet = require('helmet');
 const { errors } = require('celebrate');
 const cookieParser = require('cookie-parser');
 const cors = require('cors');
+const limiter = require('./utils/rateLimiter');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 
 const errorHandler = require('./middlewares/errorHandlers');
@@ -15,20 +15,15 @@ const router = require('./routes/index');
 const { PORT = 3000 } = process.env;
 const app = express();
 
-mongoose.connect('mongodb://localhost:27017/bitfilmsdb', {
+const { BD_NAME, NODE_ENV } = process.env;
+
+mongoose.connect(`mongodb://localhost:27017/${NODE_ENV === 'production' ? BD_NAME : 'local'}`, {
   useNewUrlParser: true,
   family: 4,
 });
 
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 100, // Limit each IP to 100 requests per `window` (here, per 15 minutes)
-  standardHeaders: true, // Return rate limit info in the `RateLimit-*` headers
-  legacyHeaders: false, // Disable the `X-RateLimit-*` headers
-});
-
 const corsOptions = {
-  // origin: 'http://localhost:3000',
+  // origin: 'http://localhost:3001',
   origin: 'https://tati-tati.nomoredomains.xyz',
   credentials: true,
 };
